veinmind-sensitive/requirements.txt
chardet==4.0.0
click==8.0.3
python_magic==0.4.25
pytoml==0.1.21
veinmind==1.0.2
